requirements.txt
python-chess>=1.999
requests>=2.31.0
pydantic>=2.0.0
pyyaml>=6.0
aiohttp>=3.9.0
